feat: Improve dark theme and fix archived tab
This commit addresses several issues related to the dark theme and the functionality of the "archived" tab.

The styling of the tabs in the dark theme has been improved to provide a better user experience. The selected tab now has a lighter grey background, and the font color is white on hover and when active. The border color of the tabs has been made consistent with the primary color of the application in both light and dark themes.

The "archived" tab, which was previously not working, has been fixed.